refactor: remove unneeded sections of files
