+++ src/airflow/dags/configs/dag.yaml
@@ -14,5 +14,4 @@
     - "my_variant_index"
     - "my_variant_annotation"
     - "my_gene_index"
-- id: "my_finngen"
 - id: "my_ukbiobank"
